+++ core/src/sn/DesignCast.ts
@@ -6,7 +6,7 @@
 ** ***** END LICENSE BLOCK ***** */
 
 import {HArg, IPropParser} from './CmnInterface';
-import {uint, int, CmnLib, argChk_Boolean, getFn} from './CmnLib';
+import {uint, int, CmnLib, argChk_Boolean, argChk_Num, getFn, addStyle} from './CmnLib';
 import {SysBase} from './SysBase';
 import {ScriptIterator} from './ScriptIterator';
 import {HPage} from './LayerMng';
@@ -18,8 +18,8 @@ import {Button} from './Button';
 import {GrpLayer} from './GrpLayer';
 import {Config} from './Config';
 
-import {Application, Rectangle, Text, Sprite} from 'pixi.js';
-import interact from 'interactjs';
+import {Application, Rectangle, Text, Sprite, Point} from 'pixi.js';
+import Moveable from 'moveable';
 
 export class DesignCast {
 	private		static	divDesignRoot: HTMLDivElement;
@@ -28,8 +28,9 @@ export class DesignCast {
 	protected	static	prpPrs		: IPropParser;
 	private		static	alzTagArg	: AnalyzeTagArg;
 	private		static	cfg			: Config;
-	static	init(appPixi: Application, sys: SysBase, scrItr: ScriptIterator, prpPrs: IPropParser, alzTagArg: AnalyzeTagArg, cfg: Config) {
-		appPixi.view.insertAdjacentHTML('beforebegin', `<div id="${this.ID_DESIGNMODE}" style="width: ${CmnLib.stageW}px; height: ${CmnLib.stageH}px; background: rgba(0,0,0,0); position: absolute; touch-action: none; user-select: none; display: none;"></div>`);
+	private		static	hPages		: HPage;
+	static	init(appPixi: Application, sys: SysBase, scrItr: ScriptIterator, prpPrs: IPropParser, alzTagArg: AnalyzeTagArg, cfg: Config, hPages: HPage) {
+		appPixi.view.insertAdjacentHTML('beforebegin', `<div id="${DesignCast.ID_DESIGNMODE}" style="width: ${CmnLib.stageW *CmnLib.cvsScale}px; height: ${CmnLib.stageH *CmnLib.cvsScale}px; background: rgba(0,0,0,0); position: absolute; touch-action: none; user-select: none; display: none;"></div>`);
 		DesignCast.divDesignRoot = document.getElementById(DesignCast.ID_DESIGNMODE) as HTMLDivElement;
 
 		DesignCast.sys = sys;
@@ -37,11 +38,50 @@ export class DesignCast {
 		DesignCast.prpPrs = prpPrs;
 		DesignCast.alzTagArg = alzTagArg;
 		DesignCast.cfg = cfg;
+		DesignCast.hPages = hPages;
+
+		addStyle(`
+.sn_design_cast {
+	position: absolute; touch-action: none; user-select: none;
+	opacity: 0.6; border-radius: 8px;
+}
+
+.sn_design_cast.drag_border {
+	line-height: 1.8;
+	border: dashed 5px #333;
+}
+`);
+		// :before	クロスカーソル.svg
+		// :after	トリミングアイコン.svg
+	}
+	static	cvsResizeDesign() {
+		const s = DesignCast.divDesignRoot.style;
+		s.width = `${CmnLib.stageW *CmnLib.cvsScale}px`;
+		s.height= `${CmnLib.stageH *CmnLib.cvsScale}px`;
 	}
 
 
-	hArg	: HArg	= {};
 	constructor(readonly bg_col: string, readonly isLay = false) {}
+
+	gethArg(): HArg {return this.hArg}
+	protected hArg	: HArg	= {};
+	protected id_tag	= '';
+	sethArg(hArg: HArg): void {
+		if (! this.id_tag) this.id_tag = hArg[':id_tag'] ?? '';
+		this.hArg = hArg;
+
+		const id_dc = hArg[':id_dc'] ?? this.id_tag;
+		DesignCast.id2gdc[id_dc] = this;
+
+		const layer = this.hArg.layer ?? '';
+		this.fncLay = (! this.parent && ! this.child && layer)
+		? ()=> {
+			const f = DesignCast.hPages[layer].fore;
+			this.lx = f.x *CmnLib.cvsScale;
+			this.ly = f.y *CmnLib.cvsScale;
+		}
+		: ()=> {};
+	}
 
 	getRect() {return Rectangle.EMPTY}
 	getPosArg(_x: number, _y: number): {[name: string]: any} {return {}};
@@ -52,118 +92,191 @@ export class DesignCast {
 	child?	: DesignCast;
 	parent?	: DesignCast;
 
-
 	private	static	readonly	ID_DESIGNMODE = 'DesignMode';
 	private	static	idDesignCast	= 0;
 	private	static	id2gdc: {[idDc: string]: DesignCast}	= {};
-	static	enterMode(node: string, hPages: HPage) {
-		const a = node.split('/');
-		const lay = hPages[a[0]];
-		if (! lay) return;
-
-		DesignCast.divDesignRoot.textContent = '';
+	static	enterMode() {
+		DesignCast.leaveMode();
 		DesignCast.divDesignRoot.style.display = 'inline';
+
 		DesignCast.idDesignCast = 0;
 		DesignCast.id2gdc = {};
-		if (a.length > 1) lay.fore.drawDesignCastChildren(
-			gdc=> gdc.dspDesignCast(hPages)
-		);
-		else lay.fore.drawDesignCast(gdc=> gdc.dspDesignCast(hPages));
 	}
+	private	static	aDC: DesignCast[] = [];
 	static	leaveMode() {
 		DesignCast.divDesignRoot.textContent = '';
 		DesignCast.divDesignRoot.style.display = 'none';
-	}
-	static	cvsResizeDesign() {
-		DesignCast.divDesignRoot.style.width = `${CmnLib.stageW}px`;
-		DesignCast.divDesignRoot.style.height= `${CmnLib.stageH}px`;
+		DesignCast.aDC.forEach(v=> {
+			v.d = null;
+			v.mov?.destroy();
+			v.mov = null;
+		});
+		DesignCast.aDC = [];
 	}
 
 
-	dspDesignCast(hPages: HPage) {
-		const o = this.hArg;
-		const id_tag = o[':id_tag'] ?? '';
-		const id_dc = o[':id_dc'] ?? id_tag;
+	cvsResize() {this.cvsResizeBase(); this.mov?.updateRect();}
+	private	cvsResizeBase() {
+		this.fncLay();
+		if (this.d) Object.assign(this.d.style, {
+			left: `${this.lx +CmnLib.cvsScale *this.rect.x}px`,
+			top: `${this.ly +CmnLib.cvsScale *this.rect.y}px`,
+			width: `${CmnLib.cvsScale *this.rect.width}px`,
+			height: `${CmnLib.cvsScale *this.rect.height}px`,
+			transformOrigin: `${this.pivot.x}px ${this.pivot.y}px`,
+			transform: `scale(${this.scale.x}, ${this.scale.y}) rotate(${this.rotation}deg)`,
+		});
+	}
+	private fncLay = ()=> {};
+
+	protected	rect	= Rectangle.EMPTY;
+	protected	pivot	= new Point(0, 0);
+	protected	scale	= new Point(1, 1);
+	protected	rotation= 0;
+	protected	mov		: Moveable | null		= null;
+	protected	d		: HTMLDivElement | null = null;
+	private lx = 0;
+	private ly = 0;
+	dspDesignCast() {
+		const id_dc = this.hArg[':id_dc'] ?? this.id_tag;
 		DesignCast.id2gdc[id_dc] = this;
 
-		let lx = 0, ly = 0;
-		if (! this.parent && ! this.child && o.layer) {
-			const lay = hPages[o.layer].fore;
-			lx = lay.x;
-			ly = lay.y;
-		}
+		this.rect = this.getRect();
 
-		const rect = this.getRect();
-//console.log(`fn:DesignCast.ts line:55 [${o.タグ名}] id_tag(${id_tag}) id_dc:(${id_dc}) fn:${o[':path']} ln:${o[':ln']} col_s:${o[':col_s']} col_e:${o[':col_e']} idx_tkn:${o[':idx_tkn']} x:${rect.x} y:${rect.y} w:${rect.width} h:${rect.height} o:%o`, o);		// token:【${o[':token']}】
-		const d = document.createElement('div');
+//console.log(`fn:DesignCast.ts dspDesignCast() [${o.タグ名}] id_tag(${id_tag}) id_dc:(${id_dc}) fn:${o[':path']} ln:${o[':ln']} col_s:${o[':col_s']} col_e:${o[':col_e']} idx_tkn:${o[':idx_tkn']} x:${rect.x} y:${rect.y} w:${rect.width} h:${rect.height} o:%o`, o);		// token:【${o[':token']}】
+
+		const d = this.d = document.createElement('div');
 		d.id = DesignCast.ID_DESIGNMODE +'_'+ ++DesignCast.idDesignCast;
-		Object.assign(d.dataset, {idDc: id_dc, x: rect.x, y: rect.y});
-		d.setAttribute('style', `
-position: absolute; touch-action: none; user-select: none;
-left: ${lx +rect.x}px;
-top: ${ly +rect.y}px;
-height: ${rect.height}px;
-width: ${rect.width}px;
-background-color: ${this.bg_col}; opacity: 0.6; border-radius: 8px;
-`);
+		d.classList.add('sn_design_cast');
+		d.dataset.id_dc = id_dc;
+		d.style.backgroundColor = `${this.bg_col}`;
+		this.cvsResizeBase();
 		(this.parent
 			? document.querySelector(
-				`[data-id-dc="${this.parent.hArg[':id_tag'] ?? ''}"]`// 親なので
+				`[data-id_dc="${this.parent.id_tag}"]`// 親なので
 				) ?? DesignCast.divDesignRoot
 			: DesignCast.divDesignRoot
 		).appendChild(d);
 
-		const me = this;
-		interact('#'+ d.id)
-		.draggable({
-			listeners: {move (e) {
-				const t = e.target;
-				let {x=0, y=0} = t.dataset;
-				x = parseFloat(x) +e.dx;
-				y = parseFloat(y) +e.dy;
-				Object.assign(t.style, {left: `${lx +x}px`, top: `${ly +y}px`});
-				Object.assign(t.dataset, {x, y});
-				const ix = int(x), iy = int(y);
-				me.delayChgCast(
-					{...me.getPosArg(ix, iy), ':id_tag': id_tag},
-					()=> me.setPos(ix, iy)
-				);
-			},},
-			modifiers: [
-				// keep the edges inside the parent
-				interact.modifiers.restrictRect({restriction: 'parent'}),
-			]
+		let oldX_Plus_W = 0;	// 整数化時誤差対応
+		let oldY_Plus_H = 0;
+		let isResizeLeft = false;
+		let isResizetop = false;
+		let isResizeWidth = false;
+		let isResizeHeight = false;
+		const mov = this.mov = new Moveable(document.body, {
+			target	: d,
+			// If the container is null, the position is fixed. (default: parentElement(document.body))
+//			container: document.body,
+			draggable	: true,
+			resizable	: true,
+			scalable	: true,
+//			rotatable	: true,
+/*
+			warpable: true,
+			// Enabling pinchable lets you use events that
+			// can be used in draggable, resizable, scalable, and rotateable.
+			pinchable: true, // ['resizable', 'scalable', 'rotatable']
+*/			// Resize, Scale Events at edges.
+/*			edge: false,
+			throttleDrag: 0,
+			throttleResize: 0,
+			throttleScale: 0,
+			throttleRotate: 0,
+*/
 		})
-		.resizable({
-			edges: {left: false, right: true, bottom: true, top: false},
-			listeners: {move(e) {
-				const t = e.target;
-				let {x=0, y=0} = t.dataset;
-				x = parseFloat(x) +e.deltaRect.left;
-				y = parseFloat(y) +e.deltaRect.top;
-				const w = uint(e.rect.width);
-				const h = uint(e.rect.height);
-				Object.assign(t.style, {
-					left	: `${lx +x}px`,
-					top		: `${ly +y}px`,
-					width	: `${w}px`,
-					height	: `${h}px`,
-				});
-				Object.assign(t.dataset, {x, y});
-				me.delayChgCast(
-					{...me.getSizeArg(w, h), ':id_tag': id_tag},
-					()=> me.setSize(w, h)
-				);
-			},},
-			modifiers: [
-				// keep the edges inside the parent
-				interact.modifiers.restrictEdges({outer: 'parent'}),
-				// minimum size
-				interact.modifiers.restrictSize({min: {width: 40, height: 40}}),
-			],
+		// draggable
+		.on('drag', ({delta})=> {
+			this.rect.x += delta[0];
+			this.rect.y += delta[1];
+			this.cvsResizeBase();
+		/*
+			d.style.left = `${left}px`;	// これだと、replaceTokenでガクッとなる
+			d.style.top = `${top}px`;
+		*/
+		}).on('dragEnd', ()=> {
+		/*
+			const r = mov.getRect();	// これだと、replaceTokenでガクッとなる
+			fncLay();
+			this.rect.x = r.left /CmnLib.cvsScale -lx;
+			this.rect.y = r.top /CmnLib.cvsScale  -ly;
+		*/
+			const ix = int(this.rect.x), iy = int(this.rect.y);
+			this.setPos(ix, iy);
+			DesignCast.sys.send2Dbg('_changeCast', {
+				...this.getPosArg(ix, iy),
+				':id_tag': this.id_tag,
+			});
 		})
-		.on('hold', ()=> DesignCast.sys.send2Dbg('_focusScript', o));
+		// resizable
+		.on('resizeStart', ()=> {
+			const ix = int(this.rect.x), iy = int(this.rect.y);
+			const iw = uint(this.rect.width), ih = uint(this.rect.height);
+			isResizeLeft = false;
+			isResizetop = false;
+			oldX_Plus_W = ix + iw;
+			oldY_Plus_H = iy + ih;
+			isResizeWidth = false;	// wをノータッチ時にhがずれる対策
+			isResizeHeight = false;	// hをノータッチ時にwがずれる対策
+		})
+		.on('resize', ({width, height, direction, dist})=> {
+			if (direction[0] < 0) {
+				isResizeLeft = true;
+				d.style.left = `${this.lx +CmnLib.cvsScale *this.rect.x -dist[0]}px`;
+				isResizeWidth = true;
+			}
+			else if (direction[0] > 0) isResizeWidth = true;
 
+			if (direction[1] < 0) {
+				isResizetop = true;
+				d.style.top = `${this.ly +CmnLib.cvsScale *this.rect.y -dist[1]}px`;
+				isResizeHeight = true;
+			}
+			else if (direction[1] > 0) isResizeHeight = true;
+
+			d.style.width = `${width}px`;
+			d.style.height = `${height}px`;
+		})
+		.on('resizeEnd', ()=> {
+			const r = mov.getRect()
+			this.fncLay();
+			if (isResizeWidth) this.rect.width  = r.width  /CmnLib.cvsScale;
+			if (isResizeHeight) this.rect.height = r.height /CmnLib.cvsScale;
+			const iw = uint(this.rect.width), ih = uint(this.rect.height);
+			this.setSize(iw, ih);
+
+			let ix = 0, iy = 0;
+			if (isResizeLeft) {	// 整数化処理で誤差が大なのでこの対応
+				ix = oldX_Plus_W -iw;	// 右端基準で新横幅戻る
+				this.rect.x = ix;
+			}
+			else ix = int(this.rect.x);	// 変化なしにて初期値
+			if (isResizetop) {
+				iy = oldY_Plus_H -ih;	// 下端基準で新縦幅上がる
+				this.rect.y = iy;
+			}
+			else iy = int(this.rect.y);	// 変化なしにて初期値
+			this.setPos(ix, iy);
+
+			DesignCast.sys.send2Dbg('_changeCast', {
+				...this.getPosArg(ix, iy),
+				...this.getSizeArg(iw, ih),
+				':id_tag': this.id_tag,
+			});
+		})
+		// rotatable
+		.on('rotate', ({transform})=> {
+			d.style.transform = transform;
+		}).on('rotateEnd', ()=> {
+			DesignCast.sys.send2Dbg('_changeCast', {
+				rotate: 0,
+//				rotate: this.mov	// TODO: 作りかけ
+				':id_tag': this.id_tag,
+			});
+		});
+		DesignCast.aDC.push(this);
+
+//		.on('hold', ()=> DesignCast.sys.send2Dbg('_focusScript', o));
 /*
 		.on('tap', e=> {
 console.log(`fn:DesignCast.ts line:162 tap btn:${e.button}`);
@@ -180,28 +293,24 @@ console.log(`fn:DesignCast.ts line:163 doubletap`);
 
 
 		// ドラッグ＆ドロップ関連
-		d.addEventListener('dragenter', ()=>
-			d.classList.add('sn_design_cast_border')
-		);
+		d.addEventListener('dragenter', ()=> d.classList.add('drag_border'));
 		d.addEventListener('dragover', e=> {
 			e.stopPropagation();
 			e.preventDefault();
 			if (e.dataTransfer) e.dataTransfer.dropEffect = 'copy';
 		});
-		d.addEventListener('dragleave', ()=>
-			d.classList.remove('sn_design_cast_border')
-		);
+		d.addEventListener('dragleave', ()=> d.classList.remove('drag_border'));
 		d.addEventListener('drop', e=> {
 			e.stopPropagation();
 			e.preventDefault();
-			d.classList.remove('sn_design_cast_border');
+			d.classList.remove('drag_border');
 			const dt = e.dataTransfer;
-			if (! dt || dt.files.length == 0) return;	// VSCodeからは 0
+			if (! dt || dt.files.length === 0) return;	// VSCodeからは 0
 
 			try {
 				const f = dt.files[0];
 				DesignCast.sys.send2Dbg('_dropFile', {
-					':id_tag'	: id_tag,
+					':id_tag': this.id_tag,
 					fn	: getFn(f.name),
 					url	: DesignCast.scrItr.cnvPath4Dbg(
 						DesignCast.cfg.searchPath(f.name, Config.EXT_SPRITE)
@@ -210,36 +319,20 @@ console.log(`fn:DesignCast.ts line:163 doubletap`);
 			} catch {}
 		});
 	}
-	static	readonly	class_def = `
-.sn_design_cast_border {
-	line-height: 1.8;
-	border: dashed 5px #333;
-}
-`;	// クラス追加は TxtLayer.ts でやってもらってる
-	// 遅延で遊びを作る
-	private tidDelay	:  NodeJS.Timer | null	= null;
-	private	delayChgCast(o: any, onSend: ()=> void) {
-		if (this.tidDelay) clearTimeout(this.tidDelay);
-		this.tidDelay = setTimeout(()=> {
-//console.log(`fn:DesignCast.ts line:175 *** delayChgCast:%o`, o);
-			onSend();
-			DesignCast.sys.send2Dbg('_changeCast', o);
-		}, 500);
-	}
 
 
-	static replaceToken(o: any, hPages: HPage) {
+	static replaceToken(o: any) {
 		const id_tag = o[':id_tag'];
 		const id_dc = id_tag;
-		const gdc = this.id2gdc[id_dc];
-		const d = <HTMLDivElement>document.querySelector(`div[data-id-dc='${id_dc}']`);
-		if (! d || ! gdc) throw `_replaceToken 存在しないidDc【${id_dc}】です`;
+		const gdc = DesignCast.id2gdc[id_dc];
+		const d = <HTMLDivElement>document.querySelector(`div[data-id_dc='${id_dc}']`);
+		if (! d || ! gdc) return;
 
 		// 青四角移動変更をスクリプト反映したレス（Undoや手変更でも呼ばれる）
 		// 内部スクリプト更新
 		const token = o[':token'];
 		DesignCast.scrItr.replace(o[':idx_tkn'], token);
-//console.log(`fn:DesignCast.ts line:193 RES id_tag(${id_tag}) ${gdc.type} o:%o`,o);
+//console.log(`fn:DesignCast.ts replaceToken id_tag(${id_tag}) ${gdc.type} o:%o`,o);
 
 		// 実ボタン・青四角も移動（Undoや手入力変更時）
 		const e = REG_TAG.exec(token);
@@ -254,25 +347,31 @@ console.log(`fn:DesignCast.ts line:163 doubletap`);
 			const y = int(DesignCast.prpPrs.getValAmpersand(p.top?.val  ?? p.y?.val ?? '0'));
 			if (isNaN(x) || isNaN(y)) DebugMng.myTrace(`widthかheightが数値ではありません\n(fn:${o[':path'].slice(13)} ln:${o[':ln']})\n${token}`, 'F');
 
-			let lx = 0, ly = 0;
-			if (! gdc.isLay && gdc.hArg.layer) {
-				const lay = hPages[gdc.hArg.layer].fore;
-				lx = lay.x;
-				ly = lay.y;
-			}
+			gdc.rect.x = x;
+			gdc.rect.y = y;
 			gdc.setPos(x, y);
-			Object.assign(d.style, {left: `${lx +x}px`, top: `${ly +y}px`});
-			Object.assign(d.dataset, {x, y});
 		}
 		if ('width' in p || 'height' in p) {
 			const w = int(DesignCast.prpPrs.getValAmpersand(p.width.val ??'0'));
 			const h = int(DesignCast.prpPrs.getValAmpersand(p.height.val??'0'));
 			if (isNaN(w) || isNaN(h)) DebugMng.myTrace(`widthかheightが数値ではありません\n(fn:${o[':path'].slice(13)} ln:${o[':ln']})\n${token}`, 'F');
 
+			gdc.rect.width = w;
+			gdc.rect.height = h;
 			gdc.setSize(w, h);
-			Object.assign(d.style, {width: `${w}px`, height: `${h}px`,});
 		}
+
+		gdc.pivot.set(
+			Number(DesignCast.prpPrs.getValAmpersand(p.pivot_x?.val ?? '0')),
+			Number(DesignCast.prpPrs.getValAmpersand(p.pivot_y?.val ?? '0')),
+		);
+		gdc.scale.set(
+			Number(DesignCast.prpPrs.getValAmpersand(p.scale_x?.val ?? '1')),
+			Number(DesignCast.prpPrs.getValAmpersand(p.scale_y?.val ?? '1')),
+		);
+		gdc.rotation = Number(DesignCast.prpPrs.getValAmpersand(p.rotation?.val ?? '0'));
 		gdc.setOther(p);
+		gdc.cvsResize();
 	}
 }
 
@@ -294,9 +393,12 @@ export class GrpLayDesignCast extends DesignCast {
 			const fn = DesignCast.prpPrs.getValAmpersand(hPrm.fn?.val ?? '');
 			this.gl.lay({fn});
 		}
+
+		this.spLay.pivot	= this.pivot;
+		this.spLay.scale	= this.scale;
+		this.spLay.angle	= this.rotation;	// angleにセット
 	}
 }
-
 
 // 文字レイヤ
 export class TxtLayDesignCast extends DesignCast {
@@ -316,7 +418,13 @@ export class TxtLayDesignCast extends DesignCast {
 	setSize(w: number, h: number) {
 		this.ts.lay(this.getSizeArg(w, h));
 	}
-	setOther(hPrm: HPRM) {this.child?.setOther(hPrm);}
+	setOther(hPrm: HPRM) {
+		this.child?.setOther(hPrm);
+
+		this.spLay.pivot	= this.pivot;
+		this.spLay.scale	= this.scale;
+		this.spLay.angle	= this.rotation;	// angleにセット
+	}
 }
 // 文字レイヤ・パディング
 export class TxtLayPadDesignCast extends DesignCast {
@@ -341,9 +449,8 @@ export class TxtLayPadDesignCast extends DesignCast {
 	setPos(pl: number, pt: number) {this.ts.lay(this.getPosArg(pl, pt));}
 	setSize(w: number, h: number) {this.ts.lay(this.getSizeArg(w, h));}
 	setOther(hPrm: HPRM) {
-		const id_tag = this.hArg[':id_tag'] ?? '';
-		const id_dc = this.hArg[':id_dc'] ?? id_tag;
-		const d = <HTMLDivElement>document.querySelector(`div[data-id-dc='${id_dc}']`);
+		const id_dc = this.hArg[':id_dc'] ?? this.id_tag;
+		const d = <HTMLDivElement>document.querySelector(`div[data-id_dc='${id_dc}']`);
 		if ('pl' in hPrm || 'pt' in hPrm) {
 			const pl = int(DesignCast.prpPrs.getValAmpersand(hPrm.pl?.val ?? '0'));
 			const pt = int(DesignCast.prpPrs.getValAmpersand(hPrm.pt?.val ?? '0'));
@@ -364,15 +471,21 @@ export class TxtLayPadDesignCast extends DesignCast {
 export class TxtBtnDesignCast extends DesignCast {
 	constructor(private readonly btn: Button, readonly hArg: HArg, private readonly txt: Text) {
 		super('#e92');
-		this.hArg = hArg;
-		if (! argChk_Boolean(hArg, 'design', true)) {
+		if (! argChk_Boolean(hArg, 'design', true)) {	// hint用
 			this.setPos = ()=> {};
 			this.setSize = ()=> {};
 		}
+
+		this.pivot.x	= argChk_Num(hArg, 'pivot_x', this.pivot.x);
+		this.pivot.y	= argChk_Num(hArg, 'pivot_y', this.pivot.y);
+		this.scale.x	= argChk_Num(hArg, 'scale_x', this.scale.x);
+		this.scale.y	= argChk_Num(hArg, 'scale_y', this.scale.y);
+		this.rotation	= argChk_Num(hArg, 'rotation', this.rotation);
+		this.sethArg(hArg);
 	}
 	getRect() {return new Rectangle(this.btn.x, this.btn.y, this.txt.width, this.txt.height)}
-	getPosArg(left: number, top: number) {return {left, top,}}
-	getSizeArg(w: number, h: number) {return {width: w, height: h,}}
+	getPosArg(left: number, top: number) {return {left, top}}
+	getSizeArg(width: number, height: number) {return {width, height}}
 	setPos(x: number, y: number) {this.btn.x = x; this.btn.y = y;}
 	setSize(w: number, h: number) {this.txt.width = w; this.txt.height = h;}
 	setOther(hPrm: HPRM) {
@@ -380,6 +493,10 @@ export class TxtBtnDesignCast extends DesignCast {
 			const b_pic = DesignCast.prpPrs.getValAmpersand(hPrm.b_pic?.val ?? '');
 			this.btn.update_b_pic(b_pic, this.txt);
 		}
+
+		this.btn.pivot	= this.pivot;
+		this.btn.scale	= this.scale;
+		this.btn.angle	= this.rotation;	// angleにセット
 	}
 }
 
@@ -387,7 +504,12 @@ export class TxtBtnDesignCast extends DesignCast {
 export class PicBtnDesignCast extends DesignCast {
 	constructor(private readonly btn: Button, readonly hArg: HArg) {
 		super('#e92');
-		this.hArg = hArg;
+		this.pivot.x	= argChk_Num(hArg, 'pivot_x', this.pivot.x);
+		this.pivot.y	= argChk_Num(hArg, 'pivot_y', this.pivot.y);
+		this.scale.x	= argChk_Num(hArg, 'scale_x', this.scale.x);
+		this.scale.y	= argChk_Num(hArg, 'scale_y', this.scale.y);
+		this.rotation	= argChk_Num(hArg, 'rotation', this.rotation);
+		this.sethArg(hArg);
 	}
 
 	private	sp: Sprite;
@@ -403,5 +525,9 @@ export class PicBtnDesignCast extends DesignCast {
 			const pic = DesignCast.prpPrs.getValAmpersand(hPrm.pic?.val ?? '');
 			this.btn.update_pic(pic, this.sp);
 		}
+
+		this.btn.pivot	= this.pivot;
+		this.btn.scale	= this.scale;
+		this.btn.angle	= this.rotation;	// angleにセット
 	}
 }